References

Clock

To record changes of array automatically, the array must refer to a time variable, which defined as a clock. There are different types of clock for continuous time ContinuousClock or discrete time DiscreteClock.

RecordedArrays.AbstractClock — Type

AbstractClock{T<:Real}

Supertype of clocks with time of type T. Clocks are used to record changes of time. A clock can be iterate by for loop and break when reach to stop.

RecordedArrays.DiscreteClock — Type

DiscreteClock{T, I<:AbstractVector{T}} <: AbstractClock{T}
DiscreteClock([start], indexset)
DiscreteClock(stop)

A clock for discrete-time process, it's indexset must be a non-empty AbstractVector. If the start is not specified, the first item of indexset will be poped and set as start. During iteration, the current time will be updated automatically and returned as iteration item. When the iteration finnished without break, init!(c) wil be applied. DiscreteClock(stop) will create a clock with start=0 and indexset=Base.OneTo(stop)

Examples

julia> c = DiscreteClock(0:3);

julia> now(c)
0

julia> [(t, now(c)) for t in c]
3-element Vector{Tuple{Int64, Int64}}:
 (1, 1)
 (2, 2)
 (3, 3)

julia> now(c)
0

julia> c = DiscreteClock(3); # similar to DiscreteClock(0:3)

julia> (now(c), collect(c))
(0, [1, 2, 3])

RecordedArrays.ContinuousClock — Type

ContinuousClock{T, I<:Union{Nothing, DiscreteClock}} <: AbstractClock{T}
ContinuousClock(stop, [start=zero(stop)]; [max_epoch=nothing])

A clock for continuous-time process. Unlike the DiscreteClock, during iteration, the current time will not be update automatically, but update by increase! manually. Besides the epoch of current iteration instead of current time will be returned. If the max_epoch is specified, the iteration will break when epoch reach to the max_epoch, even now(c) < limit(c), and break in this way the init!(c) will not be applied.

Examples

julia> c = ContinuousClock(3.0; max_epoch=2);

julia> for epoch in c
           increase!(c, 1)
           println(now(c), '	', epoch)
       end
1.0	1
2.0	2

julia> for epoch in c
           increase!(c, 1)
           println(now(c), '	', epoch)
       end
3.0	1

julia> for epoch in c
           increase!(c, 1)
           println(now(c), '	', epoch)
       end
1.0	1
2.0	2

RecordedArrays.now — Function

now(c::AbstractClock)

Return current time of clock c.

RecordedArrays.limit — Function

limit(c::AbstractClock)

Return the limit of clock c. For a ContinuousClock c, the max time might larger than limit(c).

RecordedArrays.init! — Function

init!(c::AbstractClock)

Update current time to start time.

RecordedArrays.increase! — Function

increase!(c::ContinuousClock, t::Real)

Update current time of clock c to now(c) + t.

Recorded Arrays

Create a recorded array with a clock and a array, then all changes will be recorded automatically. There are two types of recorded array: StaticRArray for array whose values of elements never change but new element will be pushed, DynamicRArray for array whose values of elements will change.

RecordedArrays.AbstractRArray — Type

AbstractRArray{V,T,N} <: AbstractArray{V,N}

Supertype of recorded N-dimensional arrays with elements of type V and time of type T", whose changes will be recorded automatically. AbstractRArray is a subtype of AbstractArray, so array operations like getindex, setindex!, length, size, mathematical operations like +, -, *, / and broadcast on a recorded array A works same as its current state state(A).

Note

Avoid to edit recorded arrays out of loop, because clocks will initial automatically during loop.

RecordedArrays.state — Function

state(A::AbstractRArray{V,T,N}) -> Array{V,N}

Get current state of recorded array A.

RecordedArrays.StaticRArray — Type

StaticRArray{V,T,N} <: AbstractRecord{V,T,N}

Record type to record changes of arrays whose elements never change but insert or delete can be created by StaticRArray(t::AbstractClock, xs...) where xs are abstract arrays to be recorded.

Implemented statical arrays:

StaticRVector

Note

Elements of StaticRArray A can be delete by deleteat!(A,i), whose value after deletion is 0.

Examples

julia> c = DiscreteClock(3);

julia> v = StaticRArray(c, [0, 1, 2])
recorded 3-element Vector{Int64}:
 0
 1
 2

julia> for epoch in c
           push!(v, epoch+2) # push a element
           deleteat!(v, 1)   # delete a element
       end

julia> v # there are still three element now
recorded 3-element Vector{Int64}:
 3
 4
 5

julia> records(v)[6] # but six element are recorded
Record Entries
t: 2-element Vector{Int64}:
 3
 3
v: 2-element Vector{Int64}:
 5
 5

julia> gettime(records(v)[1], 2)[1] # element after deletion is 0
0

RecordedArrays.DynamicRArray — Type

DynamicRArray{V,T,N} <: AbstractRecord{V,T,N}

Recorded array whose elements change overtime can be created by DynamicRArray(t::AbstractClock, xs...) where xs are abstract arrays or numbers (or called scalar) to be recorded.

Implemented dynamic arrays:

DynamicRScalar

DynamicRVector

Note

For a recorded dynamical scalar S, use S[1] = v to change its value instead of S = v.

Examples

julia> c = DiscreteClock(3);

julia> s, v = DynamicRArray(c, 0, [0, 1]);

julia> s # scalar
recorded 0

julia> v # vector
recorded 2-element Vector{Int64}:
 0
 1

julia> for epoch in c
           s[1] += 1
           v[1] += 1
       end

julia> s
recorded 3

julia> v
recorded 2-element Vector{Int64}:
 3
 1

julia> records(s)[1]
Record Entries
t: 4-element Vector{Int64}:
 0
 1
 2
 3
v: 4-element Vector{Int64}:
 0
 1
 2
 3

julia> records(v)[1]
Record Entries
t: 4-element Vector{Int64}:
 0
 1
 2
 3
v: 4-element Vector{Int64}:
 0
 1
 2
 3

julia> records(v)[2]
Record Entries
t: 1-element Vector{Int64}:
 0
v: 1-element Vector{Int64}:
 1

View Records

You can't access the recorded entries of your recorded array A directly but you can access it by create a Records with r = records(A) and get entries by r[i].

RecordedArrays.records — Function

records(A::AbstractRArray)

Create a Records with RecordedArray A.

RecordedArrays.Records — Type

Records{T<:AbstractRArray}

Contain a RecordedArray, whose elements are records of each element of given array.

RecordedArrays.AbstractEntries — Type

AbstractEntries{V,T<:Real}

Supertype of entries, which store changes of specified variable(s) of type V with time of type T.

RecordedArrays.SingleEntries — Type

SingleEntries{V,T} <: AbstractEntries{V,T}

Type to store changes of a specified variable of type V with time of type T, element of Records.

RecordedArrays.UnionEntries — Type

UnionEntries{V,T,N} <: AbstractEntries{T,V}

Type store changes of N variables of type V with time of type T, created by unione.

RecordedArrays.gettime — Function

gettime(e::AbstractEntries{V}, t::Real) -> V
gettime(e::AbstractEntries{V}, ts) -> Vector{V}

Get the value(s) of e at time t, If t is not in ts(e), return value at time ts(e)[i] where ts(e)[i] < t < ts(e)[i+1]. If a iterator of time ts is given, return the value of each time in ts.

Note

ts must be monotonically increasing.

Examples

julia> c = DiscreteClock(5);

julia> d = DynamicRArray(c, 0);

julia> s = StaticRArray(c, [0]);

julia> for t in c
           d[1] += 1
           push!(s, t)
           t >= 3 && deleteat!(s, 1)
       end

julia> ed = records(d)[1]
Record Entries
t: 6-element Vector{Int64}:
 0
 1
 2
 3
 4
 5
v: 6-element Vector{Int64}:
 0
 1
 2
 3
 4
 5

julia> es = records(s)[2]
Record Entries
t: 2-element Vector{Int64}:
 1
 4
v: 2-element Vector{Int64}:
 1
 1

julia> gettime(ed, 1.5)
1

julia> gettime(ed, [5, 6])
2-element Vector{Int64}:
 5
 5

julia> gettime(es, [2, 5])
2-element Vector{Int64}:
 1
 0

RecordedArrays.unione — Function

unione(es::AbstractEntries...)
unione(es::Vector{<:AbstractEntries})
unione(r::Records)

Construct the union of given entries es. union(r) construct union all of elements of Records r.

Examples

julia> c = DiscreteClock(3);

julia> v = DynamicRArray(c, [1, 1, 1]);

julia> for t in c
           v[t] = 2
       end

julia> ea, eb, ec = records(v);

julia> ue = unione(ea, eb)
Record Entries
t: 3-element Vector{Int64}:
 0
 1
 2
v: 3×2 Matrix{Int64}:
 1  1
 2  1
 2  2

julia> ue = unione(ue, ec)
Record Entries
t: 4-element Vector{Int64}:
 0
 1
 2
 3
v: 4×3 Matrix{Int64}:
 1  1  1
 2  1  1
 2  2  1
 2  2  2

julia> unione(records(v))
Record Entries
t: 4-element Vector{Int64}:
 0
 1
 2
 3
v: 4×3 Matrix{Int64}:
 1  1  1
 2  1  1
 2  2  1
 2  2  2

RecordedArrays.ts — Function

ts(e::AbstractEntries{V,T}) -> Vector{T}

Get time entries of given e.

RecordedArrays.vs — Function

vs(e::AbstractEntries{V,T}) -> VecOrMat{V}

Get value entries of given e.

RecordedArrays.tspan — Function

tspan(e::AbstractEntries{V,T}) -> T

Get last time of given e.

RecordedArrays.toseries — Function

toseries(e::AbstractEntries)

Convert e to the form accepted by plot of Plots.jl.