References

Clock

RecordedArrays.AbstractClock — Type

AbstractClock{T<:Real}

Supertype of clocks with time of type T.

RecordedArrays.DiscreteClock — Type

DiscreteClock([start], timelist)
DiscreteClock(stop)

Clock for discrete-time process, time of which muse be increased with given step and can't be updated manually. The timelist must be a non-empty and monotonically increasing AbstractVector. If the start is not specified, the first item of timelist will be deleted and set as start. During iteration, the current time will be updated automatically and returned as iteration item. When the iteration finished without break, init! will be applied. DiscreteClock(stop) will create a clock with start=0 and timelist=Base.OneTo(stop)

Examples

julia> c = DiscreteClock(0:3);

julia> currenttime(c)
0

julia> [(t, currenttime(c)) for t in c]
3-element Vector{Tuple{Int64, Int64}}:
 (1, 1)
 (2, 2)
 (3, 3)

julia> currenttime(c)
0

julia> c = DiscreteClock(3); # similar to DiscreteClock(0:3)

julia> (currenttime(c), collect(c))
(0, [1, 2, 3])

RecordedArrays.ContinuousClock — Type

ContinuousClock{T, I<:Union{Nothing, DiscreteClock}} <: AbstractClock{T}
ContinuousClock(stop, [start=zero(stop)]; [max_epoch=nothing])

A clock for continuous-time process. Unlike the DiscreteClock, during iteration, the current time will not be update automatically, but update by increase! manually. Besides the epoch of current iteration instead of current time will be returned. If the max_epoch is specified, the iteration will break when epoch reach to the max_epoch, even currenttime(c) < limit(c), and break in this way the init!(c) will not be applied.

Examples

julia> c = ContinuousClock(3.0; max_epoch=2);

julia> for epoch in c
           increase!(c, 1)
           println(currenttime(c), '	', epoch)
       end
1.0	1
2.0	2

julia> for epoch in c
           increase!(c, 1)
           println(currenttime(c), '	', epoch)
       end
3.0	1

julia> for epoch in c
           increase!(c, 1)
           println(currenttime(c), '	', epoch)
       end
1.0	1
2.0	2

RecordedArrays.currenttime — Function

currenttime(c::AbstractClock)

Return current time of clock c.

RecordedArrays.limit — Function

limit(c::AbstractClock)

Return the limit of clock c. For a ContinuousClock c, the max time might larger than limit(c).

RecordedArrays.start — Function

start(c::AbstractClock)

Return the start time of clock c.

RecordedArrays.init! — Function

init!(c::AbstractClock)

Update current time to start time.

RecordedArrays.increase! — Function

increase!(c::ContinuousClock, t::Real)

Update current time of clock c to currenttime(c) + t.

Recorded Arrays

RecordedArrays.AbstractRArray — Type

AbstractRArray{V,T,N} <: AbstractArray{V,N}

Supertype of recorded N-dimensional arrays with elements of type V and time of type T", whose changes will be recorded automatically. AbstractRArray is a subtype of AbstractArray, so array operations like getindex, setindex!, length, size, mathematical operations like +, -, *, / and broadcast on a recorded array A works same as its current state state(A).

Note

Avoid to edit recorded arrays out of loop, because clocks will initial automatically during loop.

RecordedArrays.StaticRArray — Type

StaticRArray{V,T,N} <: AbstractRecord{V,T,N}

Record type to record changes of arrays whose elements never change but insert or delete can be created by StaticRArray(t::AbstractClock, xs...) where xs are abstract arrays to be recorded.

Implemented statical arrays:

StaticRVector

Note

Elements of StaticRArray A can be delete by deleteat!(A,i), whose value after deletion is 0.

Examples

julia> c = DiscreteClock(3);

julia> v = StaticRArray(c, [0, 1, 2])
3-element StaticRVector{Int64, Int64, DiscreteClock{Int64, Base.OneTo{Int64}}}:
 0
 1
 2

julia> for epoch in c
           push!(v, epoch+2) # push a element
           deleteat!(v, 1)   # delete a element
       end

julia> v # there are still three element now
3-element StaticRVector{Int64, Int64, DiscreteClock{Int64, Base.OneTo{Int64}}}:
 3
 4
 5

julia> record(v)[6] # but six element are recorded
Record Entry
t: 2-element Vector{Int64}:
 3
 3
v: 2-element Vector{Int64}:
 5
 5

julia> gettime(record(v)[1], 2)[1] # element after deletion is 0
0

RecordedArrays.DynamicRArray — Type

DynamicRArray{V,T,N} <: AbstractRecord{V,T,N}

Recorded array whose elements change overtime can be created by DynamicRArray(t::AbstractClock, xs...) where xs are abstract arrays or numbers (or called scalar) to be recorded.

Implemented dynamic arrays:

DynamicRScalar

DynamicRVector

Note

For a recorded dynamical scalar S, use S[1] = v to change its value instead of S = v.

Examples

julia> c = DiscreteClock(3);

julia> s, v = DynamicRArray(c, 0, [0, 1]);

julia> s # scalar
0-dimensional DynamicRScalar{Int64, Int64, DiscreteClock{Int64, Base.OneTo{Int64}}}:
0

julia> v # vector
2-element DynamicRVector{Int64, Int64, DiscreteClock{Int64, Base.OneTo{Int64}}}:
 0
 1

julia> for epoch in c
           s[1] += 1
           v[1] += 1
       end

julia> s
0-dimensional DynamicRScalar{Int64, Int64, DiscreteClock{Int64, Base.OneTo{Int64}}}:
3

julia> v
2-element DynamicRVector{Int64, Int64, DiscreteClock{Int64, Base.OneTo{Int64}}}:
 3
 1

julia> record(s)[1]
Record Entry
t: 4-element Vector{Int64}:
 0
 1
 2
 3
v: 4-element Vector{Int64}:
 0
 1
 2
 3

julia> record(v)[1]
Record Entry
t: 4-element Vector{Int64}:
 0
 1
 2
 3
v: 4-element Vector{Int64}:
 0
 1
 2
 3

julia> record(v)[2]
Record Entry
t: 1-element Vector{Int64}:
 0
v: 1-element Vector{Int64}:
 1

RecordedArrays.state — Function

state(A::AbstractRArray{V,T,N}) -> Array{V,N}

Get current state of recorded array A. API for mathematical operations.

RecordedArrays.setclock — Function

setclock!(A::AbstractRArray, c::AbstractClock)

Non-mutating setclock for A. It will create a deepcopy of A besides the clock field, which will be assigned to c.

View Record

RecordedArrays.Record — Type

Record{T<:AbstractRArray}

Contain a RecordedArray, whose elements are record of each element of given array.

RecordedArrays.record — Function

record(A::AbstractRArray) -> Record

Create a Record with RecordedArray A.

RecordedArrays.AbstractEntry — Type

AbstractEntry{V,T<:Real}

Supertype of entry, which store changes of specified variable(s) of type V with time of type T.

RecordedArrays.SingleEntry — Type

SingleEntry{V,T} <: AbstractEntry{V,T}

Type to store changes of a specified variable of type V with time of type T, element of Record.

RecordedArrays.UnionEntry — Type

UnionEntry{V,T,N} <: AbstractEntry{T,V}

Type store changes of N variables of type V with time of type T, created by unione.

RecordedArrays.gettime — Function

gettime(e::Union{Record,AbstractEntry}, t::Real)
gettime(e::Union{Record,AbstractEntry}, ts)

Get the value(s) of e at time t, If t is not in ts(e), return value at time ts(e)[i] where ts(e)[i] < t < ts(e)[i+1]. If a iterator of time ts is given, return the value of each time in ts.

Note

ts must be monotonically increasing.

Examples

julia> c = DiscreteClock(5);

julia> d = DynamicRArray(c, 0);

julia> s = StaticRArray(c, [0]);

julia> for t in c
           d[1] += 1
           push!(s, t)
           t >= 3 && deleteat!(s, 1)
       end

julia> ed = record(d)[1]
Record Entry
t: 6-element Vector{Int64}:
 0
 1
 2
 3
 4
 5
v: 6-element Vector{Int64}:
 0
 1
 2
 3
 4
 5

julia> es = record(s)[2]
Record Entry
t: 2-element Vector{Int64}:
 1
 4
v: 2-element Vector{Int64}:
 1
 1

julia> gettime(ed, 1.5)
1

julia> gettime(ed, [5, 6])
2-element Vector{Int64}:
 5
 5

julia> gettime(es, [2, 5])
2-element Vector{Int64}:
 1
 0

RecordedArrays.unione — Function

unione(es::AbstractEntry...)
unione(es::Vector{<:AbstractEntry})
unione(r::Record)

Construct the union of given entry es. union(r) construct union all of elements of Record r.

Examples

julia> c = DiscreteClock(3);

julia> v = DynamicRArray(c, [1, 1, 1]);

julia> for t in c
           v[t] = 2
       end

julia> ea, eb, ec = record(v);

julia> ue = unione(ea, eb)
Record Entry
t: 3-element Vector{Int64}:
 0
 1
 2
v: 3×2 Matrix{Int64}:
 1  1
 2  1
 2  2

julia> ue = unione(ue, ec)
Record Entry
t: 4-element Vector{Int64}:
 0
 1
 2
 3
v: 4×3 Matrix{Int64}:
 1  1  1
 2  1  1
 2  2  1
 2  2  2

julia> unione(record(v))
Record Entry
t: 4-element Vector{Int64}:
 0
 1
 2
 3
v: 4×3 Matrix{Int64}:
 1  1  1
 2  1  1
 2  2  1
 2  2  2

RecordedArrays.getts — Function

getts(e::AbstractEntry{V,T}) -> Vector{T}

Get time entry of given e.

RecordedArrays.getvs — Function

getvs(e::AbstractEntry{V,T}) -> VecOrMat{V}

Get value entry of given e.

RecordedArrays.toseries — Function

toseries(e::AbstractEntry)

Convert e to the form accepted by plot of Plots.jl.

RecordedArrays.tspan — Function

tspan(e::AbstractEntry{V,T}) -> T

Get last time of given e.

RecordedArrays.selectrecs — Function

selectrecs([f,] r::Union{Record,AbstractEntry}, [ts], [T0], [inds...])

Select and process values in r, return a tuple of values:

f is a function for process data at each time which should accpect n parameters where n is the number of index and return a tuple. If f is not given, data will not be processed.

ts should be an iterator of real number, this function will get values at each t in ts. If ts is not given, ts will be a vector contains each time given value changed.

T0 is a constant, if T0 is given, this function will return ts as first element of returned tuple.

inds are indices, like (i1, i2, ..., in), only the values at given indices will be selected and processed. If inds is not given, all values will be selected and processed.

vars can also be unpacked, like selectrecs(r, f, ts, T0, inds...).

Example

julia> c = DiscreteClock(10); # create a clock

julia> pos = DynamicRArray(c, [0, 0]); # create a rarray

julia> for t in c # do some changes
           if  t % 2 == 0
               pos[1] += 1
           else
               pos[2] -= 1
           end
       end

julia> r = record(pos); # create a record

julia> selectrecs(r) # select without vars will return all value
([0, 0, 1, 1, 2, 2, 3, 3, 4, 4, 5], [0, -1, -1, -2, -2, -3, -3, -4, -4, -5, -5])

julia> selectrecs(r, T0, 1, 2) # select with T0 and indices
([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10], [0, 0, 1, 1, 2, 2, 3, 3, 4, 4, 5], [0, -1, -1, -2, -2, -3, -3, -4, -4, -5, -5])

julia> selectrecs(r, 0:2:10, T0, 1, 2) # select with ts and indices
(0:2:10, [0, 1, 2, 3, 4, 5], [0, -1, -2, -3, -4, -5])

julia> f(t, x, y) = t, x + y; # define a process function

julia> selectrecs(f, r, T0, 1, 2) # select with a function and indices
([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10], [0, -1, 0, -1, 0, -1, 0, -1, 0, -1, 0])

julia> selectrecs(r, T0, 1, 2) do t, x, y # or by the do block directly
           t, x + y
       end
([0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10], [0, -1, 0, -1, 0, -1, 0, -1, 0, -1, 0])